requirements.txt
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
firebase-admin==6.2.0
Flask==2.3.2
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.16
waitress==2.1.2
Werkzeug==2.3.6
